Increase compatibility with increased zoom
Previously, when the user was at a zoom higher than 100%, the form would
clip out at the bottom due to overflow being set to hidden. Overflow was
set to auto on div right-form. Additionally, a margin was added to the
same div to prevent the scrollbar from clipping with the curved borders.
Finally, the background color of the main-container was changed to match
right-form for looks.