Create a table using div and p elements
The idea behind this attempt would be to make styling and javascript
implementation to be easier. However, styling seems to be a lot harder
than with the proper table HTML elements so far, so I will switch to
table elements and give that a shot. This is a backup in case it doesn't
work.