Add styling for invalid input elements
The border will change size and will go from red to purple. Note that
the change in color is not very clear, so it would be good to add other
ways to convey whether the form is valid or not.

More evident color changes (like red to green or blue) was avoided
because these colors do not mesh well with the style of the page.