Make buttons for lifepath choices with styling
Previously, the lifepath choices were just table data (td) elements.
Hence, when hovered over, the cursor would change to an I-cursor due to
the text. With the implementation of these buttons, the text can no
longer be highlighted.

Secondly, the buttons received some interactive styling when they are
hovered for user feedback.